requirements.txt
request
httpx
beautifulsoup4
lxml
PyJWT
colorama
tqdm